neon: Move models and interfaces to models
