+++ packages/app/lib/apps.dart
@@ -1,3 +1,4 @@
+import 'package:neon/models.dart';
 import 'package:neon/platform.dart';
 import 'package:neon/utils.dart';
 import 'package:neon_files/neon_files.dart';
