--- packages/neon/neon/lib/interfaces.dart
@@ -1,1 +1,0 @@
-export 'package:neon/src/interfaces/notifications.dart';
