+++ packages/neon/neon/lib/neon.dart
@@ -8,9 +8,9 @@ import 'package:neon/src/blocs/first_launch.dart';
 import 'package:neon/src/blocs/next_push.dart';
 import 'package:neon/src/blocs/push_notifications.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
+import 'package:neon/src/models/branding.dart';
 import 'package:neon/src/platform/platform.dart';
-import 'package:neon/src/utils/app_implementation.dart';
-import 'package:neon/src/utils/branding.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/request_manager.dart';
 import 'package:package_info_plus/package_info_plus.dart';
