+++ packages/neon/neon/lib/src/app.dart
@@ -8,13 +8,13 @@ import 'package:neon/l10n/localizations.dart';
 import 'package:neon/src/bloc/result_builder.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/capabilities.dart';
-import 'package:neon/src/interfaces/notifications.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
+import 'package:neon/src/models/notifications_interface.dart';
 import 'package:neon/src/models/push_notification.dart';
 import 'package:neon/src/platform/platform.dart';
 import 'package:neon/src/router.dart';
 import 'package:neon/src/settings/widgets/option_builder.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/utils/global.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/localizations.dart';
