+++ packages/neon/neon/lib/src/blocs/accounts.dart
@@ -8,10 +8,10 @@ import 'package:neon/src/blocs/capabilities.dart';
 import 'package:neon/src/blocs/user_details.dart';
 import 'package:neon/src/blocs/user_statuses.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
 import 'package:neon/src/platform/platform.dart';
 import 'package:neon/src/settings/models/storage.dart';
 import 'package:neon/src/utils/account_options.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/request_manager.dart';
 import 'package:rxdart/rxdart.dart';
