+++ packages/neon/neon/lib/src/blocs/apps.dart
@@ -5,10 +5,10 @@ import 'package:neon/src/bloc/bloc.dart';
 import 'package:neon/src/bloc/result.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/capabilities.dart';
-import 'package:neon/src/interfaces/notifications.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
+import 'package:neon/src/models/notifications_interface.dart';
 import 'package:neon/src/settings/models/nextcloud_app_options.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/utils/request_manager.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:provider/provider.dart';
