+++ packages/neon/neon/lib/src/pages/home.dart
@@ -6,8 +6,8 @@ import 'package:neon/src/bloc/result_builder.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/apps.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
 import 'package:neon/src/settings/widgets/option_builder.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/global_options.dart' as global_options;
 import 'package:neon/src/utils/global_popups.dart';
