+++ packages/neon/neon/lib/src/pages/login.dart
@@ -3,9 +3,9 @@ import 'package:neon/l10n/localizations.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/login.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/branding.dart';
 import 'package:neon/src/platform/platform.dart';
 import 'package:neon/src/router.dart';
-import 'package:neon/src/utils/branding.dart';
 import 'package:neon/src/utils/validators.dart';
 import 'package:neon/src/widgets/exception.dart';
 import 'package:neon/src/widgets/linear_progress_indicator.dart';
