+++ packages/neon/neon/lib/src/pages/nextcloud_app_settings.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:neon/l10n/localizations.dart';
+import 'package:neon/src/models/app_implementation.dart';
 import 'package:neon/src/settings/models/select_option.dart';
 import 'package:neon/src/settings/models/toggle_option.dart';
 import 'package:neon/src/settings/widgets/checkbox_settings_tile.dart';
 import 'package:neon/src/settings/widgets/dropdown_button_settings_tile.dart';
 import 'package:neon/src/settings/widgets/settings_category.dart';
 import 'package:neon/src/settings/widgets/settings_list.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/utils/confirmation_dialog.dart';
 
 class NextcloudAppSettingsPage extends StatelessWidget {
