+++ packages/neon/neon/lib/src/utils/settings_export_helper.dart
@@ -1,6 +1,6 @@
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
 import 'package:neon/src/settings/models/option.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/utils/global_options.dart';
 
 class SettingsExportHelper {
