+++ packages/neon/neon/lib/src/widgets/app_bar.dart
@@ -6,10 +6,10 @@ import 'package:neon/l10n/localizations.dart';
 import 'package:neon/src/bloc/result_builder.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/apps.dart';
-import 'package:neon/src/interfaces/notifications.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
+import 'package:neon/src/models/notifications_interface.dart';
 import 'package:neon/src/router.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/widgets/app_implementation_icon.dart';
 import 'package:neon/src/widgets/exception.dart';
 import 'package:neon/src/widgets/linear_progress_indicator.dart';
