+++ packages/neon/neon/lib/src/widgets/app_implementation_icon.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:neon/src/utils/app_implementation.dart';
+import 'package:neon/src/models/app_implementation.dart';
 import 'package:neon/src/widgets/user_avatar.dart';
 
 class NeonAppImplementationIcon extends StatelessWidget {
