+++ packages/neon/neon/lib/src/widgets/drawer.dart
@@ -8,8 +8,8 @@ import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/blocs/apps.dart';
 import 'package:neon/src/blocs/capabilities.dart';
 import 'package:neon/src/models/account.dart';
+import 'package:neon/src/models/app_implementation.dart';
 import 'package:neon/src/router.dart';
-import 'package:neon/src/utils/app_implementation.dart';
 import 'package:neon/src/widgets/account_tile.dart';
 import 'package:neon/src/widgets/cached_image.dart';
 import 'package:neon/src/widgets/drawer_destination.dart';
